Migrate errors to seperate file
