+++ backend/src/errors/Errors.ts
@@ -1,0 +1,34 @@
+export class BadRequestError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = "BadRequestError";
+    }
+}
+
+export class ValidationError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = "ValidationError";
+    }
+}
+
+export class DatabaseError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = "DatabaseError";
+    }
+}
+
+export class NotFoundError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = "NotFoundError";
+    }
+}
+
+export class ConflictError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = "ConflictError";
+    }
+}
