+++ backend/src/errors/index.ts
@@ -1,40 +1,6 @@
 import { Response } from "express";
 import { dateFormat, timeFormat } from "../utils";
-
-export class BadRequestError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = "BadRequestError";
-    }
-}
-
-export class ValidationError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = "ValidationError";
-    }
-}
-
-export class DatabaseError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = "DatabaseError";
-    }
-}
-
-export class NotFoundError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = "NotFoundError";
-    }
-}
-
-export class ConflictError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = "ConflictError";
-    }
-}
+import { BadRequestError, ConflictError, DatabaseError, NotFoundError, ValidationError } from "./Errors";
 
 export function handleErrorResponse(error: Error, res: Response) {
     switch (true) {
